@@ -1,16 +1,17 @@
 import { NgZone } from '@angular/core';
-import { createStore, Reducer, Action, Store } from 'redux';
+import { createStore, Reducer, Action, AnyAction, Store } from 'redux';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/combineLatest';
+import { Observable } from 'rxjs';
+import { combineLatest, filter } from 'rxjs/operators';
 
 import { NgRedux } from './ng-redux';
 import { RootStore } from './root-store';
 import { select } from '../decorators/select';
 
-class MockNgZone {
-  run = (fn: Function) => fn();
+class MockNgZone extends NgZone {
+  run<T>(fn: (...args: any[]) => T): T {
+    return fn() as T;
+  }
 }
 
 type PayloadAction = Action & { payload?: string | number };
@@ -23,10 +24,10 @@ describe('NgRedux Observable Store', () => {
   }
 
   let defaultState: IAppState;
-  let rootReducer: Reducer<IAppState>;
+  let rootReducer: Reducer<IAppState, AnyAction>;
   let store: Store<IAppState>;
   let ngRedux: NgRedux<IAppState>;
-  const mockNgZone = new MockNgZone() as NgZone;
+  const mockNgZone = new MockNgZone({ enableLongStackTrace: false }) as NgZone;
 
   beforeEach(() => {
     defaultState = {
@@ -62,26 +63,26 @@ describe('NgRedux Observable Store', () => {
   });
 
   it('should get the initial state', done =>
-    ngRedux.select().subscribe((state: IAppState) => {
+    ngRedux.select<any>().subscribe((state: IAppState) => {
       expect(state.foo).toEqual('bar');
       expect(state.baz).toEqual(-1);
       done();
     }));
 
   it('should accept a keyname for a selector', done =>
-    ngRedux.select('foo').subscribe(stateSlice => {
+    ngRedux.select<any>('foo').subscribe(stateSlice => {
       expect(stateSlice).toEqual('bar');
       done();
     }));
 
   it('should not trigger selector if that slice of state wasnt changed', () => {
-    let fooData;
+    let fooData = '';
 
     const spy = jasmine.createSpy('spy').and.callFake((foo: string) => {
       fooData = foo;
     });
 
-    const foo$ = ngRedux.select('foo').subscribe(spy);
+    const foo$ = ngRedux.select<any>('foo').subscribe(spy);
 
     expect(spy.calls.count()).toEqual(1);
     ngRedux.dispatch({ type: 'UPDATE_BAR', payload: 0 });
@@ -95,7 +96,7 @@ describe('NgRedux Observable Store', () => {
   });
 
   it('should not trigger a selector if the action payload is the same', () => {
-    let fooData;
+    let fooData = '';
     const spy = jasmine.createSpy('spy').and.callFake((foo: string) => {
       fooData = foo;
     });
@@ -111,7 +112,7 @@ describe('NgRedux Observable Store', () => {
   });
 
   it('should not call sub if the result of the function is the same', () => {
-    let fooData;
+    let fooData = '';
     const spy = jasmine.createSpy('spy').and.callFake((foo: string) => {
       fooData = foo;
     });
@@ -223,7 +224,7 @@ describe('NgRedux Observable Store', () => {
 
     const someService = new SomeService();
     someService.foo$
-      .combineLatest(someService.bar$, someService.baz$)
+      .pipe(combineLatest(someService.bar$, someService.baz$))
       .subscribe(([foo, bar, baz]) => {
         expect(foo).toEqual('bar');
         expect(bar).toEqual('foo');
@@ -242,9 +243,9 @@ describe('Chained actions in subscriptions', () => {
   }
 
   let defaultState: IAppState;
-  let rootReducer: Reducer<IAppState>;
+  let rootReducer: Reducer<IAppState, AnyAction>;
   let ngRedux: NgRedux<IAppState>;
-  const mockNgZone = new MockNgZone() as NgZone;
+  const mockNgZone = new MockNgZone({ enableLongStackTrace: false }) as NgZone;
 
   const doSearch = (word: string) =>
     ngRedux.dispatch({ type: 'SEARCH', payload: word });
@@ -276,7 +277,7 @@ describe('Chained actions in subscriptions', () => {
     it(`length sub should be called twice`, () => {
       const keyword$ = ngRedux.select(n => n.keyword);
       let keyword = '';
-      let length;
+      let length = 0;
       const length$ = ngRedux.select(n => n.keywordLength);
       const lengthSpy = jasmine
         .createSpy('lengthSpy')
@@ -284,7 +285,7 @@ describe('Chained actions in subscriptions', () => {
       let lenSub;
       let keywordSub;
 
-      keywordSub = keyword$.filter(n => n !== '').subscribe(n => {
+      keywordSub = keyword$.pipe(filter(n => n !== '')).subscribe(n => {
         keyword = n;
         doFetch(n);
       });
@@ -307,7 +308,7 @@ describe('Chained actions in subscriptions', () => {
     it(`second sub should get most current state value`, () => {
       const keyword$ = ngRedux.select(n => n.keyword);
       let keyword = '';
-      let length;
+      let length = 0;
       const length$ = ngRedux.select(n => n.keywordLength);
       const lengthSpy = jasmine
         .createSpy('lengthSpy')
@@ -315,7 +316,7 @@ describe('Chained actions in subscriptions', () => {
       let lenSub;
       let keywordSub;
 
-      keywordSub = keyword$.filter(n => n !== '').subscribe(n => {
+      keywordSub = keyword$.pipe(filter(n => n !== '')).subscribe(n => {
         keyword = n;
         doFetch(n);
       });
@@ -339,7 +340,7 @@ describe('Chained actions in subscriptions', () => {
     it(`length sub should be called twice`, () => {
       const keyword$ = ngRedux.select(n => n.keyword);
       let keyword = '';
-      let length;
+      let length = 0;
       const length$ = ngRedux.select(n => n.keywordLength);
       const lengthSpy = jasmine
         .createSpy('lengthSpy')
@@ -348,7 +349,7 @@ describe('Chained actions in subscriptions', () => {
       let keywordSub;
 
       lenSub = length$.subscribe(lengthSpy);
-      keywordSub = keyword$.filter(n => n !== '').subscribe(n => {
+      keywordSub = keyword$.pipe(filter(n => n !== '')).subscribe(n => {
         keyword = n;
         doFetch(n);
       });
@@ -369,7 +370,7 @@ describe('Chained actions in subscriptions', () => {
     it(`first sub should get most current state value`, () => {
       const keyword$ = ngRedux.select(n => n.keyword);
       let keyword = '';
-      let length;
+      let length = 0;
       const length$ = ngRedux.select(n => n.keywordLength);
       const lengthSpy = jasmine
         .createSpy('lengthSpy')
@@ -378,7 +379,7 @@ describe('Chained actions in subscriptions', () => {
       let keywordSub;
 
       lenSub = length$.subscribe(lengthSpy);
-      keywordSub = keyword$.filter(n => n !== '').subscribe(n => {
+      keywordSub = keyword$.pipe(filter(n => n !== '')).subscribe(n => {
         keyword = n;
         doFetch(n);
       });
