@@ -1,10 +1,8 @@
 import { NgZone } from '@angular/core';
 import { Action } from 'redux';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/toArray';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs';
+import { map, take, toArray } from 'rxjs/operators';
 
 import { NgRedux } from '../components/ng-redux';
 import { RootStore } from '../components/root-store';
@@ -46,8 +44,7 @@ describe('Select decorators', () => {
         const mockInstance = new MockClass();
 
         mockInstance.baz
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-1, 1]),
             undefined,
@@ -64,8 +61,7 @@ describe('Select decorators', () => {
         const mockInstance = new MockClass();
 
         mockInstance.baz$
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-1, 4]),
             undefined,
@@ -84,8 +80,7 @@ describe('Select decorators', () => {
         const mockInstance = new MockClass();
 
         mockInstance.obs$
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-1, 3]),
             undefined,
@@ -104,8 +99,7 @@ describe('Select decorators', () => {
         const mockInstance = new MockClass();
 
         mockInstance.obs$
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-2, 10]),
             undefined,
@@ -125,8 +119,7 @@ describe('Select decorators', () => {
       it('should only trigger next when comparator returns true', done => {
         const mockInstance = new MockClass();
         mockInstance.baz$
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-1, 2]),
             undefined,
@@ -144,7 +137,7 @@ describe('Select decorators', () => {
         }
 
         const mockInstance = new LocalMockClass();
-        mockInstance.baz$.take(3).subscribe(undefined, undefined, done);
+        mockInstance.baz$.pipe(take(3)).subscribe(undefined, undefined, done);
 
         ngRedux.dispatch({ type: 'nvm', payload: 1 });
         ngRedux.dispatch({ type: 'nvm', payload: 2 });
@@ -156,7 +149,8 @@ describe('Select decorators', () => {
   });
 
   describe('@select$', () => {
-    const transformer = (baz$: Observable<number>) => baz$.map(baz => 2 * baz);
+    const transformer = (baz$: Observable<number>) =>
+      baz$.pipe(map(baz => 2 * baz));
 
     it('applies a transformer to the observable', done => {
       class MockClass {
@@ -166,8 +160,7 @@ describe('Select decorators', () => {
       const mockInstance = new MockClass();
 
       mockInstance.baz$
-        .take(2)
-        .toArray()
+        .pipe(take(2), toArray())
         .subscribe(values => expect(values).toEqual([-2, 10]), undefined, done);
       ngRedux.dispatch({ type: 'nvm', payload: 5 });
     });
@@ -182,8 +175,7 @@ describe('Select decorators', () => {
       it('should only trigger next when the comparator returns true', done => {
         const mockInstance = new MockClass();
         mockInstance.baz$
-          .take(2)
-          .toArray()
+          .pipe(take(2), toArray())
           .subscribe(
             values => expect(values).toEqual([-2, 2]),
             undefined,
@@ -201,7 +193,7 @@ describe('Select decorators', () => {
         }
 
         const mockInstance = new SpyClass();
-        mockInstance.baz$.take(3).subscribe(undefined, undefined, done);
+        mockInstance.baz$.pipe(take(3)).subscribe(undefined, undefined, done);
 
         ngRedux.dispatch({ type: 'nvm', payload: 1 });
         ngRedux.dispatch({ type: 'nvm', payload: 2 });
