@@ -1,3 +1,3 @@
-﻿var typedoc = typedoc || {};
+var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"64":"Function","128":"Class","256":"Interface","1024":"Property","2048":"Method","65536":"Type literal","4194304":"Type alias"},"rows":[{"id":0,"kind":4194304,"name":"Comparator","url":"globals.html#comparator","classes":"tsd-kind-type-alias"},{"id":1,"kind":65536,"name":"__type","url":"globals.html#comparator.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"Comparator"},{"id":2,"kind":4194304,"name":"Transformer","url":"globals.html#transformer","classes":"tsd-kind-type-alias"},{"id":3,"kind":65536,"name":"__type","url":"globals.html#transformer.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"Transformer"},{"id":4,"kind":4194304,"name":"PropertySelector","url":"globals.html#propertyselector","classes":"tsd-kind-type-alias"},{"id":5,"kind":4194304,"name":"PathSelector","url":"globals.html#pathselector","classes":"tsd-kind-type-alias"},{"id":6,"kind":4194304,"name":"FunctionSelector","url":"globals.html#functionselector","classes":"tsd-kind-type-alias"},{"id":7,"kind":65536,"name":"__type","url":"globals.html#functionselector.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"FunctionSelector"},{"id":8,"kind":4194304,"name":"Selector","url":"globals.html#selector","classes":"tsd-kind-type-alias"},{"id":9,"kind":256,"name":"ObservableStore","url":"interfaces/observablestore.html","classes":"tsd-kind-interface tsd-has-type-parameter"},{"id":10,"kind":1024,"name":"select","url":"interfaces/observablestore.html#select","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"ObservableStore"},{"id":11,"kind":65536,"name":"__type","url":"interfaces/observablestore.html#select.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"ObservableStore.select"},{"id":12,"kind":1024,"name":"configureSubStore","url":"interfaces/observablestore.html#configuresubstore","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"ObservableStore"},{"id":13,"kind":65536,"name":"__type","url":"interfaces/observablestore.html#configuresubstore.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"ObservableStore.configureSubStore"},{"id":14,"kind":1024,"name":"dispatch","url":"interfaces/observablestore.html#dispatch","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":15,"kind":2048,"name":"getState","url":"interfaces/observablestore.html#getstate","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":16,"kind":2048,"name":"subscribe","url":"interfaces/observablestore.html#subscribe","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":17,"kind":2048,"name":"replaceReducer","url":"interfaces/observablestore.html#replacereducer","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":18,"kind":128,"name":"NgRedux","url":"classes/ngredux.html","classes":"tsd-kind-class tsd-has-type-parameter"},{"id":19,"kind":1024,"name":"instance","url":"classes/ngredux.html#instance","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"NgRedux"},{"id":20,"kind":1024,"name":"configureStore","url":"classes/ngredux.html#configurestore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":21,"kind":65536,"name":"__type","url":"classes/ngredux.html#configurestore.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.configureStore"},{"id":22,"kind":1024,"name":"provideStore","url":"classes/ngredux.html#providestore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":23,"kind":65536,"name":"__type","url":"classes/ngredux.html#providestore.__type-3","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.provideStore"},{"id":24,"kind":1024,"name":"dispatch","url":"classes/ngredux.html#dispatch","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":25,"kind":1024,"name":"getState","url":"classes/ngredux.html#getstate","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":26,"kind":65536,"name":"__type","url":"classes/ngredux.html#getstate.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.getState"},{"id":27,"kind":1024,"name":"subscribe","url":"classes/ngredux.html#subscribe","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":28,"kind":65536,"name":"__type","url":"classes/ngredux.html#subscribe.__type-6","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.subscribe"},{"id":29,"kind":1024,"name":"replaceReducer","url":"classes/ngredux.html#replacereducer","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":30,"kind":65536,"name":"__type","url":"classes/ngredux.html#replacereducer.__type-4","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.replaceReducer"},{"id":31,"kind":1024,"name":"select","url":"classes/ngredux.html#select","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":32,"kind":65536,"name":"__type","url":"classes/ngredux.html#select.__type-5","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"NgRedux.select"},{"id":33,"kind":1024,"name":"configureSubStore","url":"classes/ngredux.html#configuresubstore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":34,"kind":65536,"name":"__type","url":"classes/ngredux.html#configuresubstore.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"NgRedux.configureSubStore"},{"id":35,"kind":128,"name":"DevToolsExtension","url":"classes/devtoolsextension.html","classes":"tsd-kind-class"},{"id":36,"kind":2048,"name":"enhancer","url":"classes/devtoolsextension.html#enhancer","classes":"tsd-kind-method tsd-parent-kind-class","parent":"DevToolsExtension"},{"id":37,"kind":2048,"name":"isEnabled","url":"classes/devtoolsextension.html#isenabled","classes":"tsd-kind-method tsd-parent-kind-class","parent":"DevToolsExtension"},{"id":38,"kind":64,"name":"enableFractalReducers","url":"globals.html#enablefractalreducers","classes":"tsd-kind-function"},{"id":39,"kind":256,"name":"IFractalStoreOptions","url":"interfaces/ifractalstoreoptions.html","classes":"tsd-kind-interface"},{"id":40,"kind":1024,"name":"basePathMethodName","url":"interfaces/ifractalstoreoptions.html#basepathmethodname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"IFractalStoreOptions"},{"id":41,"kind":1024,"name":"localReducer","url":"interfaces/ifractalstoreoptions.html#localreducer","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"IFractalStoreOptions"},{"id":42,"kind":64,"name":"select","url":"globals.html#select","classes":"tsd-kind-function tsd-has-type-parameter"},{"id":43,"kind":64,"name":"select$","url":"globals.html#select_","classes":"tsd-kind-function tsd-has-type-parameter"},{"id":44,"kind":64,"name":"dispatch","url":"globals.html#dispatch","classes":"tsd-kind-function"},{"id":45,"kind":64,"name":"WithSubStore","url":"globals.html#withsubstore","classes":"tsd-kind-function"},{"id":46,"kind":128,"name":"NgReduxModule","url":"classes/ngreduxmodule.html","classes":"tsd-kind-class"}]};
+            typedoc.search.data = {"kinds":{"64":"Function","128":"Class","256":"Interface","1024":"Property","2048":"Method","65536":"Type literal","4194304":"Type alias"},"rows":[{"id":0,"kind":4194304,"name":"Comparator","url":"globals.html#comparator","classes":"tsd-kind-type-alias"},{"id":1,"kind":65536,"name":"__type","url":"globals.html#comparator.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"Comparator"},{"id":2,"kind":4194304,"name":"Transformer","url":"globals.html#transformer","classes":"tsd-kind-type-alias tsd-has-type-parameter"},{"id":3,"kind":65536,"name":"__type","url":"globals.html#transformer.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"Transformer"},{"id":4,"kind":4194304,"name":"PropertySelector","url":"globals.html#propertyselector","classes":"tsd-kind-type-alias"},{"id":5,"kind":4194304,"name":"PathSelector","url":"globals.html#pathselector","classes":"tsd-kind-type-alias"},{"id":6,"kind":4194304,"name":"FunctionSelector","url":"globals.html#functionselector","classes":"tsd-kind-type-alias tsd-has-type-parameter"},{"id":7,"kind":65536,"name":"__type","url":"globals.html#functionselector.__type","classes":"tsd-kind-type-literal tsd-parent-kind-type-alias tsd-is-not-exported","parent":"FunctionSelector"},{"id":8,"kind":4194304,"name":"Selector","url":"globals.html#selector","classes":"tsd-kind-type-alias tsd-has-type-parameter"},{"id":9,"kind":256,"name":"ObservableStore","url":"interfaces/observablestore.html","classes":"tsd-kind-interface tsd-has-type-parameter"},{"id":10,"kind":1024,"name":"select","url":"interfaces/observablestore.html#select","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"ObservableStore"},{"id":11,"kind":65536,"name":"__type","url":"interfaces/observablestore.html#select.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"ObservableStore.select"},{"id":12,"kind":1024,"name":"configureSubStore","url":"interfaces/observablestore.html#configuresubstore","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"ObservableStore"},{"id":13,"kind":65536,"name":"__type","url":"interfaces/observablestore.html#configuresubstore.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"ObservableStore.configureSubStore"},{"id":14,"kind":1024,"name":"dispatch","url":"interfaces/observablestore.html#dispatch","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":15,"kind":2048,"name":"getState","url":"interfaces/observablestore.html#getstate","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":16,"kind":2048,"name":"subscribe","url":"interfaces/observablestore.html#subscribe","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":17,"kind":2048,"name":"replaceReducer","url":"interfaces/observablestore.html#replacereducer","classes":"tsd-kind-method tsd-parent-kind-interface tsd-is-inherited","parent":"ObservableStore"},{"id":18,"kind":128,"name":"NgRedux","url":"classes/ngredux.html","classes":"tsd-kind-class tsd-has-type-parameter"},{"id":19,"kind":1024,"name":"instance","url":"classes/ngredux.html#instance","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"NgRedux"},{"id":20,"kind":1024,"name":"configureStore","url":"classes/ngredux.html#configurestore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":21,"kind":65536,"name":"__type","url":"classes/ngredux.html#configurestore.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.configureStore"},{"id":22,"kind":1024,"name":"provideStore","url":"classes/ngredux.html#providestore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":23,"kind":65536,"name":"__type","url":"classes/ngredux.html#providestore.__type-3","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.provideStore"},{"id":24,"kind":1024,"name":"dispatch","url":"classes/ngredux.html#dispatch","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":25,"kind":1024,"name":"getState","url":"classes/ngredux.html#getstate","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":26,"kind":65536,"name":"__type","url":"classes/ngredux.html#getstate.__type-2","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.getState"},{"id":27,"kind":1024,"name":"subscribe","url":"classes/ngredux.html#subscribe","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":28,"kind":65536,"name":"__type","url":"classes/ngredux.html#subscribe.__type-6","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.subscribe"},{"id":29,"kind":1024,"name":"replaceReducer","url":"classes/ngredux.html#replacereducer","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":30,"kind":65536,"name":"__type","url":"classes/ngredux.html#replacereducer.__type-4","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"NgRedux.replaceReducer"},{"id":31,"kind":1024,"name":"select","url":"classes/ngredux.html#select","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":32,"kind":65536,"name":"__type","url":"classes/ngredux.html#select.__type-5","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"NgRedux.select"},{"id":33,"kind":1024,"name":"configureSubStore","url":"classes/ngredux.html#configuresubstore","classes":"tsd-kind-property tsd-parent-kind-class","parent":"NgRedux"},{"id":34,"kind":65536,"name":"__type","url":"classes/ngredux.html#configuresubstore.__type-1","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-has-type-parameter tsd-is-not-exported","parent":"NgRedux.configureSubStore"},{"id":35,"kind":128,"name":"DevToolsExtension","url":"classes/devtoolsextension.html","classes":"tsd-kind-class"},{"id":36,"kind":2048,"name":"enhancer","url":"classes/devtoolsextension.html#enhancer","classes":"tsd-kind-method tsd-parent-kind-class","parent":"DevToolsExtension"},{"id":37,"kind":2048,"name":"isEnabled","url":"classes/devtoolsextension.html#isenabled","classes":"tsd-kind-method tsd-parent-kind-class","parent":"DevToolsExtension"},{"id":38,"kind":64,"name":"enableFractalReducers","url":"globals.html#enablefractalreducers","classes":"tsd-kind-function"},{"id":39,"kind":256,"name":"IFractalStoreOptions","url":"interfaces/ifractalstoreoptions.html","classes":"tsd-kind-interface"},{"id":40,"kind":1024,"name":"basePathMethodName","url":"interfaces/ifractalstoreoptions.html#basepathmethodname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"IFractalStoreOptions"},{"id":41,"kind":1024,"name":"localReducer","url":"interfaces/ifractalstoreoptions.html#localreducer","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"IFractalStoreOptions"},{"id":42,"kind":64,"name":"select","url":"globals.html#select","classes":"tsd-kind-function tsd-has-type-parameter"},{"id":43,"kind":64,"name":"select$","url":"globals.html#select_","classes":"tsd-kind-function tsd-has-type-parameter"},{"id":44,"kind":64,"name":"dispatch","url":"globals.html#dispatch","classes":"tsd-kind-function"},{"id":45,"kind":64,"name":"WithSubStore","url":"globals.html#withsubstore","classes":"tsd-kind-function"},{"id":46,"kind":128,"name":"NgReduxModule","url":"classes/ngreduxmodule.html","classes":"tsd-kind-class"}]};
