@@ -9,12 +9,16 @@ export const setIn = (
   obj: any,
   [firstElem, ...restElems]: (string | number)[],
   value: any
-): Object => typeof (obj[ firstElem ] || {}).setIn === 'function' ? {
-  ...obj, [ firstElem ]: obj[ firstElem ].setIn(restElems, value)
-} : {
-  ...obj,
-  [firstElem]:
-    restElems.length === 0
-      ? value
-      : setIn(obj[firstElem] || {}, restElems, value),
-};
+): Object =>
+  'function' === typeof (obj[firstElem] || {}).setIn
+    ? {
+        ...obj,
+        [firstElem]: obj[firstElem].setIn(restElems, value),
+      }
+    : {
+        ...obj,
+        [firstElem]:
+          restElems.length === 0
+            ? value
+            : setIn(obj[firstElem] || {}, restElems, value),
+      };
